+++ src/App.js
@@ -1,16 +1,28 @@
+import { Route, Routes } from 'react-router-dom';
 import './App.css';
-import Calculator from './components/calculator';
-import Quote from './components/Quote';
+import Calculator from './routes/Calculator';
+import Layout from './components/Layout';
 import './css/calculator.css';
+import Home from './routes/Home';
+import Quote from './routes/Quote';
 
 function App() {
   return (
-    <div className="App">
-      <div className="main-grp">
-        <Quote />
-        <Calculator />
-      </div>
-    </div>
+    <>
+      {/* <div className="App">
+        <div className="main-grp">
+          <Quote />
+          <Calculator />
+        </div>
+      </div> */}
+      <Routes>
+        <Route path="/" element={<Layout />}>
+          <Route index element={<Home />} />
+          <Route path="/calculator" element={<Calculator />} />
+          <Route path="/quote" element={<Quote />} />
+        </Route>
+      </Routes>
+    </>
   );
 }
