+++ src/components/logic.js/calculate.js
@@ -73,7 +73,6 @@ export default function calculate(obj, buttonName) {
         operation: null,
       };
     }
-    // '=' with no operation, nothing to do
     return {};
   }
