+++ src/routes/Home.jsx
@@ -1,0 +1,12 @@
+const Home = () => (
+  <>
+    <h1>Welcome to our page!</h1>
+    <p>
+      Lorem ipsum dolor sit amet, consectetur adipisicing elit. At facere unde
+      ipsum quae quisquam sapiente provident harum repellat sit, temporibus
+      corporis consequatur distinctio vero dolorum ab sunt tempora ipsa minus?
+    </p>
+  </>
+);
+
+export default Home;
